feat: Ubuntu VM
Add Ubuntu machines, refactor vagrantfiles and fix scripts.

+++ scripts/trigger/set_ssh_config.ps1
@@ -7,8 +7,8 @@ IP of the host in ssh config file.
 Name of the host in ssh config file.
 .EXAMPLE
 $IpAddress = '192.168.121.88'
-$HostName = 'vagranthv'
-scripts/set_ssh_config.ps1 $IpAddress $HostName
+$HostName = 'fedorahv'
+scripts/trigger/set_ssh_config.ps1 $IpAddress $HostName
 #>
 [CmdletBinding()]
 param (
@@ -28,18 +28,33 @@ Host $HostName
 "@
 
 Write-Host "Adding $HostName to ssh config..."
+if (-not (Test-Path $sshConfig)) { New-Item $sshConfig -ItemType File -Force }
 if (Select-String -Pattern "HostName $IpAddress" -Path $sshConfig) {
+    # update Host if HostName entry already present
     $content = [IO.File]::ReadAllText($sshConfig) -replace "Host[^\n]+\n[^\n]+$IpAddress\n[\s\S]+?(?=(\nHost|\z))", $vagrantConfig
+    [IO.File]::WriteAllText($sshConfig, $content)
+} elseif (Select-String -Pattern "Host $HostName" -Path $sshConfig) {
+    # update HostName if Host entry already present
+    $content = [IO.File]::ReadAllText($sshConfig) -replace "Host $HostName[\s\S]+?(?=(\nHost|\z))", $vagrantConfig
     [IO.File]::WriteAllText($sshConfig, $content)
 } else {
     Add-Content -Value $vagrantConfig -Path $sshConfig
 }
 
-Write-Host 'Cleaning ssh known_hosts file...'
-if (Select-String -Pattern "^$IpAddress" -Path $knownHosts) {
-    $content = [IO.File]::ReadAllLines($knownHosts) -notmatch "^$IpAddress"
-    [IO.File]::WriteAllLines($knownHosts, $content)
+if (Test-Path $knownHosts) {
+    Write-Host 'Cleaning ssh known_hosts file...'
+    if (Select-String -Pattern "^$IpAddress" -Path $knownHosts) {
+        $content = [IO.File]::ReadAllLines($knownHosts) -notmatch "^$IpAddress"
+        [IO.File]::WriteAllLines($knownHosts, $content)
+    }
 }
 
-Write-Host 'Adding fingerprint to ssh known_hosts file...'
-ssh-keyscan $IpAddress 2>$null | Add-Content -Path $knownHosts
+if (Get-Command ssh-keyscan -ErrorAction SilentlyContinue) {
+    Write-Host 'Adding fingerprint to ssh known_hosts file...'
+    do {
+        $knownIP = ssh-keyscan $IpAddress 2>$null
+        if ($knownIP) {
+            $knownIP | Add-Content -Path $knownHosts
+        }
+    } until ($knownIP)
+}
